Renamed and merge buffer headers
