+++ engine/renderer/private/bloom_settings.cpp
@@ -3,7 +3,7 @@
 #include "graphics_context.hpp"
 #include "graphics_resources.hpp"
 #include "pipeline_builder.hpp"
-#include "resource_management/buffer_resource_manager.hpp"
+#include "resources/buffer.hpp"
 #include "settings.hpp"
 #include "vulkan_helper.hpp"
