+++ engine/renderer/public/batch_buffer.hpp
@@ -4,7 +4,7 @@
 #include "constants.hpp"
 #include "graphics_context.hpp"
 #include "graphics_resources.hpp"
-#include "resource_management/buffer_resource_manager.hpp"
+#include "resources/buffer.hpp"
 #include "single_time_commands.hpp"
 #include "vertex.hpp"
 #include "vulkan_helper.hpp"
