+++ engine/renderer/public/graphics_resources.hpp
@@ -3,10 +3,14 @@
 #include "common.hpp"
 #include <memory>
 
+namespace bb
+{
+class BufferManager;
+}
+
 class VulkanContext;
 class MeshResourceManager;
 class MaterialResourceManager;
-class BufferResourceManager;
 class SamplerResourceManager;
 class ModelResourceManager;
 class ImageResourceManager;
@@ -23,7 +27,7 @@ public:
     MeshResourceManager& GetMeshResourceManager() { return *_meshResourceManager; }
     MaterialResourceManager& GetMaterialResourceManager() { return *_materialResourceManager; }
     ImageResourceManager& GetImageResourceManager() { return *_imageResourceManager; }
-    BufferResourceManager& GetBufferResourceManager() { return *_bufferResourceManager; }
+    bb::BufferManager& GetBufferResourceManager() { return *_bufferResourceManager; }
     SamplerResourceManager& GetSamplerResourceManager() { return *_samplerResourceManager; }
     ModelResourceManager& GetModelResourceManager() { return *_modelResourceManager; }
 
@@ -34,6 +38,6 @@ private:
     std::shared_ptr<MeshResourceManager> _meshResourceManager;
     std::shared_ptr<MaterialResourceManager> _materialResourceManager;
     std::shared_ptr<ImageResourceManager> _imageResourceManager;
-    std::shared_ptr<BufferResourceManager> _bufferResourceManager;
+    std::shared_ptr<bb::BufferManager> _bufferResourceManager;
     std::shared_ptr<SamplerResourceManager> _samplerResourceManager;
 };
